* commit 'c970939deca3e3bbdbf45210c6cc8aa8a3a7b15f':
  Se modifica template para mostrar los 500 caracteres en observaciones.
  Se corrige int a decimal
  correct colors
  Fix se muestran los decimales en detalle del pedido

+++ services/pedidos-service/Omicron.Pedidos.Repository/Model/CompleteDetailOrderModel.cs
@@ -43,7 +43,7 @@ namespace Omicron.Pedidos.Entities.Model
         /// Gets or sets Code.
         /// </summary>
         /// <value>The code.</value>
-        public int? QtyPlanned { get; set; }
+        public decimal? QtyPlanned { get; set; }
 
         /// <summary>
         /// Gets or sets Code.
